+++ include/openssl/rsa.h
@@ -127,7 +127,7 @@ OPENSSL_EXPORT int RSA_encrypt(const BIGNUM *n, const BIGNUM *e,
 
 /* RSA_size returns the number of bytes in the modulus, which is also the size
  * of a signature or encrypted value using |rsa|. */
-OPENSSL_EXPORT unsigned RSA_size(const RSA *rsa);
+OPENSSL_EXPORT size_t RSA_size(const RSA *rsa);
 
 /* RSAPrivateKey_dup allocates a fresh |RSA| and copies the private key from
  * |rsa| into it. It returns the fresh |RSA| object, or NULL on error. */
